+++ CHANGES.rst
@@ -9,6 +9,9 @@ New Features:
 - Add schema to @registry listing.
   [jaroel]
 
+- Manipulate group membership in the @groups endpoint.
+  [jaroel]
+
 - Expose a user's global roles though the @users endpoint.
   [jaroel]